+++ src/agent/server.js
@@ -1,10 +1,12 @@
-import { config, defer, async, t, log } from 'azk';
+import { config, defer, async, t, log, path } from 'azk';
 import { VM  }   from 'azk/agent/vm';
 import { Unfsd } from 'azk/agent/unfsd';
 import { Balancer } from 'azk/agent/balancer';
 import { net as net_utils } from 'azk/utils';
 import { AgentStartError } from 'azk/utils/errors';
 import { Api } from 'azk/agent/api';
+
+var qfs = require('q-io/fs');
 
 var Server = {
   server: null,
@@ -101,6 +103,19 @@ var Server = {
         var key = config('agent:vm:ssh_key') + '.pub';
         var authoried = config('agent:vm:authorized_key');
         yield VM.copyFile(vm_name, key, authoried);
+
+        // Get docker keys
+        var files = ['ca.pem', 'cert.pem', 'key.pem'];
+        var origin, dest, pems = config('paths:pems');
+        yield qfs.makeTree(pems);
+
+        // Copy files
+        for (var i = 0; i < files.length; i++) {
+          n("docker_keys");
+          origin = path.join("/home/docker/.docker", files[i]);
+          dest   = path.join(pems, files[i]);
+          yield VM.copyVMFile(vm_name, origin, dest);
+        }
       };
 
       // Mark installed